requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
lxml==6.0.0
openai==1.93.0
outcome==1.3.0.post0
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
realtime==2.5.3
requests==2.32.4
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
